requirements.txt
unstructured==0.9.2
tiktoken
faiss-cpu==1.7.4
libmagic

python-magic-bin

streamlit>=1.0
python-dotenv
langchain>=0.0.150  # Or the latest version you're using
langchain-openai

playwright
selenium
